fix:Improve tests separation based on code review comments
