+++ core/control-plane/control-plane-aggregate-services/src/test/java/org/eclipse/edc/connector/controlplane/services/contractdefinition/ContractDefinitionServiceImplTest.java
@@ -29,23 +29,15 @@ import org.eclipse.edc.transaction.spi.NoopTransactionContext;
 import org.eclipse.edc.transaction.spi.TransactionContext;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.ArgumentsProvider;
-import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import java.util.UUID;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.edc.spi.query.Criterion.criterion;
 import static org.eclipse.edc.spi.result.ServiceFailure.Reason.CONFLICT;
 import static org.eclipse.edc.spi.result.ServiceFailure.Reason.NOT_FOUND;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.argThat;
@@ -210,56 +202,6 @@ class ContractDefinitionServiceImplTest {
         verify(listener, never()).updated(any());
     }
 
-    @Nested
-    class QueryValidatorIntegrationTest {
-        private final ContractDefinitionService service = new ContractDefinitionServiceImpl(store, transactionContext, observable);
-
-        @ParameterizedTest
-        @ArgumentsSource(InvalidFilters.class)
-        void search_invalidFilter(Criterion invalidFilter) {
-            var query = QuerySpec.Builder.newInstance()
-                    .filter(invalidFilter)
-                    .build();
-
-            var result = service.search(query);
-
-            assertThat(result.failed()).isTrue();
-        }
-
-        @ParameterizedTest
-        @ArgumentsSource(ValidFilters.class)
-        void search_validFilter(Criterion validFilter) {
-            var query = QuerySpec.Builder.newInstance()
-                    .filter(validFilter)
-                    .build();
-            when(store.findAll(query)).thenReturn(Stream.empty());
-
-            service.search(query);
-
-            verify(store).findAll(query);
-        }
-
-        private static class InvalidFilters implements ArgumentsProvider {
-            @Override
-            public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
-                return Stream.of(
-                        arguments(criterion("assetsSelector.leftHand", "=", "foo")), // invalid path
-                        arguments(criterion("accessPolicyId'LIKE/**/?/**/LIMIT/**/?/**/OFFSET/**/?;DROP/**/TABLE/**/test/**/--%20", "=", "%20ABC--")) //some SQL injection
-                );
-            }
-        }
-
-        private static class ValidFilters implements ArgumentsProvider {
-            @Override
-            public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
-                return Stream.of(
-                        arguments(criterion("assetsSelector.operandLeft", "=", "foo")),
-                        arguments(criterion("assetsSelector.operator", "=", "LIKE")),
-                        arguments(criterion("assetsSelector.operandRight", "=", "bar"))
-                );
-            }
-        }
-    }
 
     @NotNull
     private Predicate<ContractDefinition> hasId(String id) {
