+++ core/control-plane/control-plane-aggregate-services/src/test/java/org/eclipse/edc/connector/controlplane/services/transferprocess/TransferProcessServiceImplTest.java
@@ -29,7 +29,6 @@ import org.eclipse.edc.connector.controlplane.transfer.spi.types.command.Suspend
 import org.eclipse.edc.connector.controlplane.transfer.spi.types.command.TerminateTransferCommand;
 import org.eclipse.edc.spi.command.CommandHandlerRegistry;
 import org.eclipse.edc.spi.command.CommandResult;
-import org.eclipse.edc.spi.query.Criterion;
 import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.spi.response.StatusResult;
 import org.eclipse.edc.spi.result.Result;
@@ -42,11 +41,6 @@ import org.eclipse.edc.validator.spi.DataAddressValidatorRegistry;
 import org.eclipse.edc.validator.spi.ValidationResult;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.ArgumentsProvider;
-import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
@@ -58,7 +52,6 @@ import static org.eclipse.edc.spi.query.Criterion.criterion;
 import static org.eclipse.edc.spi.result.ServiceFailure.Reason.BAD_REQUEST;
 import static org.eclipse.edc.spi.result.ServiceFailure.Reason.NOT_FOUND;
 import static org.eclipse.edc.validator.spi.Violation.violation;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.mock;
@@ -259,56 +252,6 @@ class TransferProcessServiceImplTest {
         assertThat(result.failed()).isTrue();
         assertThat(result.getFailureMessages()).containsExactly("not found");
         verify(transactionContext).execute(any(TransactionContext.ResultTransactionBlock.class));
-    }
-
-    @Nested
-    class QueryValidatorIntegrationTest {
-        private final TransferProcessService service = new TransferProcessServiceImpl(store, manager, transactionContext,
-                dataAddressValidator, commandHandlerRegistry);
-
-        @ParameterizedTest
-        @ArgumentsSource(InvalidFilters.class)
-        void search_invalidFilter_raiseException(Criterion invalidFilter) {
-            var spec = QuerySpec.Builder.newInstance().filter(invalidFilter).build();
-
-            var result = service.search(spec);
-
-            assertThat(result.failed()).isTrue();
-        }
-
-        @ParameterizedTest
-        @ArgumentsSource(ValidFilters.class)
-        void search_validFilter(Criterion validFilter) {
-            var spec = QuerySpec.Builder.newInstance().filter(validFilter).build();
-
-            service.search(spec);
-
-            verify(store).findAll(spec);
-            verify(transactionContext).execute(any(TransactionContext.ResultTransactionBlock.class));
-        }
-
-        private static class InvalidFilters implements ArgumentsProvider {
-            @Override
-            public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
-                return Stream.of(
-                        arguments(criterion("provisionedResourceSet.resources.hastoken", "=", "true")), // wrong case
-                        arguments(criterion("resourceManifest.definitions.notexist", "=", "foobar")), // property not exist
-                        arguments(criterion("contentDataAddress.properties[*].someKey", "=", "someval")) // map types not supported
-                );
-            }
-        }
-
-        private static class ValidFilters implements ArgumentsProvider {
-            @Override
-            public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
-                return Stream.of(
-                        arguments(criterion("deprovisionedResources.provisionedResourceId", "=", "someval")),
-                        arguments(criterion("type", "=", "CONSUMER")),
-                        arguments(criterion("provisionedResourceSet.resources.hasToken", "=", "true"))
-                );
-            }
-        }
-
     }
 
     private TransferProcess transferProcess() {
